+++ packages/article-skeleton/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.72.1](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.72.0...@times-components/article-skeleton@1.72.1) (2022-03-01)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
 # [1.72.0](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.71.4...@times-components/article-skeleton@1.72.0) (2022-02-28)